backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-dotenv
alembic
boto3
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
python-jose[cryptography]
pydantic-settings
redis
google-api-python-client
google-auth
resend